Use multiple GPU

+++ Training.py
@@ -89,9 +89,9 @@ class SupervisedTrainer(BaseTrainer):
 
             X, targets, IDs = batch
             targets = targets.to(self.device)
-            predictions = self.model(X.to(self.device))
+            predictions = self.model(X.half().to(self.device))
 
-            loss = self.loss_module(predictions, targets)  # (batch_size,) loss for each sample in the batch
+            loss = self.loss_module(predictions.float(), targets)  # (batch_size,) loss for each sample in the batch
             batch_loss = torch.sum(loss)
             mean_loss = batch_loss / len(loss)  # mean loss (over samples) used for optimization
             total_loss = mean_loss
